test: update GitService tests to use specific dummy repository URL
